+++ changelog.md
@@ -2,3 +2,4 @@
 
 # 2018-08-11
 - Added minimum type of characters functionality.
+- Fixed bug to prevent user from selecting more minimum characters then password length.